feat: soporte venta a granel — stock y cantidades decimales
- Agregar campos sells_by_weight y unit al modelo Product
- Convertir stock, quantity y campos relacionados de Integer a Decimal
- Eliminar SaleUnit (redundante), unificar en Unit (PZ, KG, CO)
- Eliminar source_unit/target_unit de ProductConversion (se toma del producto)
- Mover endpoint units a GET /api/products/units/
- Crear SmartDecimalField para serializar: entero si no tiene decimales, float si sí
- Corregir mixed types (DecimalField vs IntegerField) en annotations
- Convertir quantity a Decimal en creación de venta para evitar TypeError
- Agregar .env.example, .kiro/rules.md y spec granel-sales.yaml

+++ sales/models.py
@@ -99,7 +99,7 @@ class ProductSale(models.Model):
     sale = models.ForeignKey(
         Sale, on_delete=models.CASCADE, related_name="products_sale"
     )
-    quantity = models.IntegerField()
+    quantity = models.DecimalField(max_digits=10, decimal_places=3)
     price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def get_total(self):
